+++ starfish/spots/_pixel_decoder/_base.py
@@ -1,25 +1,68 @@
-from typing import Callable, Sequence, Tuple, Union
+from abc import abstractmethod
+from typing import Callable, Sequence, Tuple, Type
 
+import click
 import numpy as np
-import xarray as xr
 
+from starfish.codebook.codebook import Codebook
 from starfish.imagestack.imagestack import ImageStack
 from starfish.intensity_table.intensity_table import IntensityTable
 from starfish.pipeline.algorithmbase import AlgorithmBase
-from starfish.types import Number, SpotAttributes
+from starfish.pipeline.pipelinecomponent import PipelineComponent
+from starfish.types import Number
 from .combine_adjacent_features import ConnectedComponentDecodingResult
 
 
+COMPONENT_NAME = "detect_pixels"
+
+
+class PixelSpotDecoder(PipelineComponent):
+    @classmethod
+    def pipeline_component_type_name(cls) -> str:
+        return COMPONENT_NAME
+
+    @classmethod
+    def _cli_run(cls, ctx, instance):
+        output = ctx.obj["output"]
+        image_stack = ctx.obj["image_stack"]
+        # TODO ambrosejcarr serialize and save ConnectedComponentDecodingResult somehow
+        intensities, ccdr = instance.run(image_stack)
+        intensities.save(output)
+
+    @staticmethod
+    @click.group(COMPONENT_NAME)
+    @click.option("-i", "--input", required=True, type=click.Path(exists=True))
+    @click.option("-o", "--output", required=True)
+    @click.option(
+        '--codebook', default=None, required=True, help=(
+            'A spaceTx spec-compliant json file that describes a three dimensional tensor '
+            'whose values are the expected intensity of a spot for each code in each imaging '
+            'round and each color channel.'
+        )
+    )
+    @click.pass_context
+    def _cli(ctx, input, output, codebook):
+        """pixel-wise spot detection and decoding"""
+        print('Detecting Spots ...')
+        ctx.obj = dict(
+            component=PixelSpotDecoder,
+            image_stack=ImageStack.from_path_or_url(input),
+            output=output,
+            codebook=Codebook.from_json(codebook),
+        )
+
+
 class PixelDecoderAlgorithmBase(AlgorithmBase):
+    @classmethod
+    def get_pipeline_component_class(cls) -> Type[PipelineComponent]:
+        return PixelSpotDecoder
+
+    @abstractmethod
     def run(
             self,
             primary_image: ImageStack,
     ) -> Tuple[IntensityTable, ConnectedComponentDecodingResult]:
         """Finds spots in an ImageStack"""
-        raise NotImplementedError()
-
-    def image_to_spots(self, data_image: Union[np.ndarray, xr.DataArray]) -> SpotAttributes:
-        """Finds spots in a 3d volume"""
         raise NotImplementedError()
 
     @staticmethod
